fix: read arrays A and B in separate loops instead of a combined loop
Previously, I was using:
for (i=0; i<n; i++) {
    scanf(%d, &a[i]);
    scanf(%d, &b[i]);
}

This caused incorrect input mapping because the problem input gives all of A first,
then all of B on the next line. Reading both in the same loop mixed up the values.

Updated the code to:
- read all elements of A in one loop
- read all elements of B in a separate loop

This correctly matches the input format and fixes the wrong answer issue.